+++ mosaico-sdk-py/src/mosaicolabs/bridges/protocols/mcaps/__init__.py
@@ -1,0 +1,3 @@
+from . import converters as converters
+from .base_converter import SchemaConverter as SchemaConverter
+from .registry import SchemaRegistry as SchemaRegistry
