+++ mosaico-sdk-py/src/mosaicolabs/bridges/ros/adapter_base.py
