+++ mosaico-sdk-py/src/mosaicolabs/bridges/ros/data_ontology/battery_state.py
