+++ mosaico-sdk-py/src/mosaicolabs/bridges/ros/loader.py
@@ -24,13 +24,13 @@ from mosaicolabs import (
     SequenceHandler,
     TopicHandler,
 )
-from mosaicolabs.bridges.ros_bridge.adapters.unmodeled import UnmodeledAdapter
+from mosaicolabs.bridges.ros.adapters.unmodeled import UnmodeledAdapter
 from mosaicolabs.enum.serialization_format import SerializationFormat
 from mosaicolabs.logging_config import get_logger
 from mosaicolabs.models.core.helpers import resolve_ontology_class
 
 from ...models.core.serializable import _compute_schema_fingerprint
-from ...protocols.converters.ros_converter import RosMsgConverter
+from ..protocols.mcaps.converters.ros_converter import RosMsgConverter
 from .adapter_base import ROSAdapterBase, RosSchemaMetadata
 from .helpers import (
     _class_name_from_ros_msgtype,
