+++ mosaico-sdk-py/src/mosaicolabs/bridges/ros/registry.py
