--- mosaico-sdk-py/src/mosaicolabs/protocols/__init__.py
@@ -1,3 +1,0 @@
-from . import converters as converters
-from .base_converter import McapConverter as McapConverter
-from .mcap_bridge import McapBridge as McapBridge
