+++ mosaico-sdk-py/src/testing/unit/models/test_schema_generation.py
@@ -45,7 +45,7 @@ from mosaicolabs import (
     Vector4d,
     Velocity,
 )
-from mosaicolabs.bridges.ros_bridge.data_ontology import (
+from mosaicolabs.bridges.ros.data_ontology import (
     BatteryState,
     FrameTransform,
     PointCloud2,
